sample_data/scripts/train.py
"""
ADVANCE Training Script
Visual Navigation with Adaptive Deep Learning
"""

import torch
import torch.nn as nn
import numpy as np
from torch.optim import Adam
from torch.utils.data import DataLoader

# Hyperparameters
LEARNING_RATE = 0.0001
BATCH_SIZE = 256
NUM_EPOCHS = 20
HIDDEN_SIZE = 512
DROPOUT = 0.1

# Note: Random seed should be set for reproducibility
# TODO: Add explicit seed setting

class VisualEncoder(nn.Module):
    """Transformer-based visual encoder."""
    
    def __init__(self, input_dim=256, hidden_dim=512, num_heads=8):
        super().__init__()
        self.embedding = nn.Linear(input_dim * input_dim * 3, hidden_dim)
        self.transformer = nn.TransformerEncoder(
            nn.TransformerEncoderLayer(d_model=hidden_dim, nhead=num_heads),
            num_layers=6
        )
        self.output = nn.Linear(hidden_dim, hidden_dim)
    
    def forward(self, x):
        x = x.flatten(start_dim=1)
        x = self.embedding(x)
        x = self.transformer(x.unsqueeze(0))
        return self.output(x.squeeze(0))


class PolicyNetwork(nn.Module):
    """LSTM-based policy network."""
    
    def __init__(self, visual_dim=512, action_dim=4, hidden_dim=256):
        super().__init__()
        self.lstm = nn.LSTM(visual_dim, hidden_dim, batch_first=True)
        self.policy_head = nn.Linear(hidden_dim, action_dim)
        self.value_head = nn.Linear(hidden_dim, 1)
    
    def forward(self, visual_features, hidden=None):
        lstm_out, hidden = self.lstm(visual_features.unsqueeze(1), hidden)
        policy = self.policy_head(lstm_out.squeeze(1))
        value = self.value_head(lstm_out.squeeze(1))
        return policy, value, hidden


class ADVANCE(nn.Module):
    """Full ADVANCE model."""
    
    def __init__(self):
        super().__init__()
        self.encoder = VisualEncoder()
        self.policy = PolicyNetwork()
    
    def forward(self, observation, hidden=None):
        visual_features = self.encoder(observation)
        return self.policy(visual_features, hidden)


def train_epoch(model, dataloader, optimizer, device):
    """Train for one epoch."""
    model.train()
    total_loss = 0
    
    for batch_idx, (observations, actions, rewards) in enumerate(dataloader):
        observations = observations.to(device)
        actions = actions.to(device)
        rewards = rewards.to(device)
        
        optimizer.zero_grad()
        
        policy, value, _ = model(observations)
        
        # Simplified PPO loss (for demonstration)
        policy_loss = -torch.mean(policy * actions)
        value_loss = torch.mean((value.squeeze() - rewards) ** 2)
        loss = policy_loss + 0.5 * value_loss
        
        loss.backward()
        optimizer.step()
        
        total_loss += loss.item()
    
    return total_loss / len(dataloader)


def evaluate(model, dataloader, device):
    """Evaluate the model."""
    model.eval()
    correct = 0
    total = 0
    
    with torch.no_grad():
        for observations, actions, _ in dataloader:
            observations = observations.to(device)
            actions = actions.to(device)
            
            policy, _, _ = model(observations)
            predictions = torch.argmax(policy, dim=1)
            targets = torch.argmax(actions, dim=1)
            
            correct += (predictions == targets).sum().item()
            total += len(predictions)
    
    return correct / total if total > 0 else 0


def main():
    """Main training function."""
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    print(f"Using device: {device}")
    
    model = ADVANCE().to(device)
    optimizer = Adam(model.parameters(), lr=LEARNING_RATE)
    
    # Training loop would go here
    # Using placeholder for demonstration
    print("Training ADVANCE model...")
    print(f"Hyperparameters: LR={LEARNING_RATE}, Batch={BATCH_SIZE}, Epochs={NUM_EPOCHS}")
    
    # Note: Actual training code would use proper data loading
    # This is a simplified demonstration


if __name__ == "__main__":
    main()
