requirements.txt
google-adk>=1.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
pylatexenc>=2.10
bibtexparser>=1.4.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
aiofiles>=23.0.0
